Split "pdfCheck.py" into multiple submodules
