+++ scanner/constants.py
@@ -1,0 +1,31 @@
+DEADLINE_DATE_STR = "2018-09-23T00:00:00+02:00"
+OUTPUT_FIELDS = [
+    "Site",
+    "File",
+    "Accessible",
+    "TotallyInaccessible",
+    "BrokenFile",
+    "TaggedTest",
+    "EmptyTextTest",
+    "ProtectedTest",
+    "TitleTest",
+    "LanguageTest",
+    "BookmarksTest",
+    "Exempt",
+    "Date",
+    "hasTitle",
+    "hasDisplayDocTitle",
+    "hasLang",
+    "InvalidLang",
+    "Form",
+    "xfa",
+    "hasBookmarks",
+    "hasXmp",
+    "PDFVersion",
+    "Creator",
+    "Producer",
+    "Pages",
+    "_log",
+    "fonts",
+    "numTxtObjects",
+]
